# openapi.action.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GitHub Agent Gateway",
    "version": "1.0.0",
    "description": "Safely manage allowlisted GitHub repositories. Preview code changes before applying them. Destructive and admin operations require exact confirmation strings and server-side feature flags."
  },
  "servers": [
    {
      "url": "https://YOUR-WORKER.workers.dev"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/repos/{owner}/{repository}/pulls": {
      "get": {
        "summary": "List pull requests",
        "operationId": "listPullRequests",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Create a pull request",
        "operationId": "createPullRequest",
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "head",
                  "base"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "head": {
                    "type": "string"
                  },
                  "base": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string"
                  },
                  "draft": {
                    "type": "boolean",
                    "default": true
                  },
                  "maintainerCanModify": {
                    "type": "boolean"
                  },
                  "changePlanId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/repos/{owner}/{repository}/pulls/{number}": {
      "get": {
        "summary": "Get a pull request",
        "operationId": "getPullRequest",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "summary": "Update a pull request",
        "operationId": "updatePullRequest",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/v1/repos/{owner}/{repository}/pulls/{number}/comments": {
      "post": {
        "summary": "Comment on a pull request",
        "operationId": "commentOnPullRequest",
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "body"
                ],
                "properties": {
                  "body": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/repos/{owner}/{repository}/pulls/{number}/reviewers": {
      "post": {
        "summary": "Request pull request reviewers",
        "operationId": "requestPullRequestReviewers",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reviewers": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "teamReviewers": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove requested pull request reviewers",
        "operationId": "removePullRequestReviewers",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reviewers": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "teamReviewers": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/repos/{owner}/{repository}/pulls/{number}/reviews": {
      "post": {
        "summary": "Create a pull request review",
        "operationId": "createPullRequestReview",
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/v1/repos/{owner}/{repository}/pulls/{number}/merge": {
      "put": {
        "summary": "Merge a pull request",
        "operationId": "mergePullRequest",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "description": "Disabled by default. Confirmation: MERGE owner/repository#number.",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "confirmation"
                ],
                "properties": {
                  "mergeMethod": {
                    "type": "string",
                    "enum": [
                      "merge",
                      "squash",
                      "rebase"
                    ]
                  },
                  "expectedHeadSha": {
                    "type": "string"
                  },
                  "commitTitle": {
                    "type": "string"
                  },
                  "commitMessage": {
                    "type": "string"
                  },
                  "confirmation": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}